UF-174 : Move of GenericPortableException and ExceptionUtilities into uberfire-extensions
